Fix flaky `test_fix_metadata_version_on_attach_part_after_restore`
Move `test_fix_metadata_version_on_attach_part_after_restore` from
`test_restore_replica` to its own test module
`test_restore_replica_metadata_version`.

The test shared a module-scoped cluster with 4 other tests that perform
10+ `SYSTEM RESTART REPLICA` operations. Each restart destroys and
recreates storage objects with background threads. In TSAN builds, the
accumulated thread churn can exhaust TSAN internal thread slot tracking,
causing a segfault in `__tsan::TraceMutexUnlock` during a system table
background merge — crashing the server and failing this unrelated test.

The test is fully independent: it creates its own `test_ttl` table on
only 2 nodes and shares no state with the other restore replica tests.
Giving it a fresh cluster avoids the accumulated TSAN thread debris.

https://s3.amazonaws.com/clickhouse-test-reports/json.html?PR=96437&sha=de1caa72d9160d3d01d29245a6f7f1ba453a46e1&name_0=PR&name_1=Integration%20tests%20%28amd_tsan%2C%206%2F6%29

Closes #88500
