+++ tests/integration/test_restore_replica/test.py
@@ -188,52 +188,3 @@ def test_restore_replica_alive_replicas(start_cluster):
 
     check_after_restoration()
     drop_tables()
-
-
-def test_fix_metadata_version_on_attach_part_after_restore(start_cluster):
-    zk = cluster.get_kazoo_client("zoo1")
-
-    node_1.query("DROP TABLE IF EXISTS test_ttl ON CLUSTER 'test_cluster' SYNC")
-    for node in [node_1, node_2]:
-        node.query(
-            """
-            CREATE TABLE test_ttl(n UInt32, d DateTime)
-            ENGINE = ReplicatedMergeTree('/clickhouse/tables/test_ttl/', '{replica}')
-            TTL d + INTERVAL 5 SECOND DELETE
-            ORDER BY n PARTITION BY n % 10
-            SETTINGS merge_with_ttl_timeout = 0, number_of_free_entries_in_pool_to_execute_mutation = 0;
-        """.format(
-                replica=node.name
-            )
-        )
-
-    # Alter metadata to increment metadata version of the table
-    node_1.query("ALTER TABLE test_ttl ADD COLUMN Added1 UInt32")
-    node_1.query("ALTER TABLE test_ttl DROP COLUMN Added1")
-
-    node_1.query("SYSTEM STOP TTL MERGES test_ttl")
-
-    # Create a part
-    node_1.query("INSERT INTO test_ttl VALUES (1, now())")
-
-    # Delete root zk metadata path for the table
-    zk_rmr_with_retries(zk, "/clickhouse/tables/test_ttl")
-
-    # Restore replicas
-    node_1.query("SYSTEM RESTART REPLICA test_ttl")
-    node_1.query("SYSTEM RESTORE REPLICA test_ttl")
-    # Disable TTL merges on node_1. We expect it to be done on node_2
-    node_1.query("SYSTEM STOP TTL MERGES test_ttl")
-
-    node_2.query("SYSTEM RESTART REPLICA test_ttl")
-    node_2.query("SYSTEM RESTORE REPLICA test_ttl")
-
-    # TTL merge should work.
-    # Before fix it is failed due to metadata version discrepancy and we will have in the log:
-    # "Source part metadata version 2 is newer then the table metadata version 0. ALTER_METADATA is still in progress"
-    assert_eq_with_retry(
-        node_2, "SELECT count() FROM test_ttl", "0\n", retry_count=60, sleep_time=1
-    )
-
-    for node in [node_1, node_2]:
-        node.query("DROP TABLE IF EXISTS test_ttl SYNC")
